Refactor message volume analysis by moving clustering logic to `utils/ofw/clusters.js`
- Introduced `analyzeRapidFireMessages` function in `clusters.js` for better organization of message clustering logic.
- Added explicit JSON file input validation in `message-volume.js`.
- Updated imports in `message-volume.js` to utilize the new clustering utility.
